Start to code SSL support
git-svn-id: svn://svn.savannah.gnu.org/myserver/trunk@386 c4c922ea-a70d-0410-937d-ec1daf91936c

+++ myserverweb/include/clientsThread.h
@@ -53,7 +53,7 @@ private:
 	void controlConnections();
 	u_long nBytesToRead;
 public:
-	const static int ID_OFFSET = 200;
+	const static u_long ID_OFFSET = 200;
 	ClientsTHREAD();
 	~ClientsTHREAD();
 	void stop();
